ce-genai-analytics/backend/requirements.txt
fastapi
uvicorn

# Database
sqlalchemy
pyodbc

# Environment variables
python-dotenv

# OpenAI
openai

# Validation
pydantic
